+++ daemon/api/router.go
@@ -96,6 +96,8 @@ func NewRouter(h *Handlers, static http.Handler) http.Handler {
 			authed.Get("/system/host", h.GetSystemHost)
 			authed.Get("/system/integrations", h.GetSystemIntegrations)
 			authed.Get("/system/spool", h.GetSystemSpool)
+			authed.Get("/system/libraries", h.GetSystemLibraries)
+			authed.Post("/system/libraries/recalc", h.RecalcSystemLibraries)
 
 			authed.Get("/discs/{id}/episodes", h.GetDiscEpisodes)